finishing phase 1

+++ resources/views/subject/index.blade.php
@@ -1,7 +1,5 @@
 @extends('layouts.app')
 @section('content')
-<link href="{{ asset('css/fab.css') }}" rel="stylesheet">
-<link href="{{ asset('css/table.css') }}" rel="stylesheet">
 <div class="card" style="margin-bottom: 100px;">
     <div class="card-body">
         <div class="card-title">
@@ -25,21 +23,16 @@
                         <td>{{ $subject->description }}</td>
                         <td class="center">
                             @if($myData->myRole == "dosen")
-                            <a href="subjectdelete/{{ $subject->id }}" class="a"><button class="action btn"
-                                    onclick="return confirm('yakin?');" id="delete">Delete</button></a>
-                            <a data-target="#assignstore-{{$subject->id}}" data-toggle="modal"
-                                class="a btn btn-warning"> Tugaskan </a>
+                            <a href="subjectdelete/{{ $subject->id }}" class="a"><button class="action btn" onclick="return confirm('yakin?');" id="delete">Delete</button></a>
+                            <a data-target="#assignstore-{{$subject->id}}" data-toggle="modal" class="a btn btn-warning"> Tugaskan </a>
                             @endif
                             @if($myData->myRole == "mahasiswa")
                             <a href="{{ $subject->file }}"><button class="action btn btn-info">Lihat Tugas</button></a>
                             @endif
-                            <a href="{{ $subject->file }}" class="a"><button class="action btn"
-                                    id="download">Lihat</button></a>
+                            <a href="{{ $subject->file }}" class="a"><button class="action btn" id="download">Lihat</button></a>
                         </td>
                     </tr>
-
-                    <div class="modal fade" id="assignstore-{{$subject->id}}" tabindex="-1" role="dialog"
-                        aria-labelledby="myModalLabel" aria-hidden="true">
+                    <div class="modal fade" id="assignstore-{{$subject->id}}" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
                         <div class="modal-dialog">
                             <div class="modal-content">
                                 <!-- Modal root -->
@@ -50,16 +43,13 @@
                                 <div class="inputs">
                                     <div class="content">
                                         <div class="col-md-12">
-                                            <form action="{{ route('create.assign') }}" method="post"
-                                                enctype="multipart/form-data">
+                                            <form action="{{ route('create.assign') }}" method="post" enctype="multipart/form-data">
                                                 @csrf
                                                 <div class="row mb-3">
-                                                    <label for="assignFile"
-                                                        class="col-md-4 col-form-label text-md-end">File
+                                                    <label for="assignFile" class="col-md-4 col-form-label text-md-end">File
                                                         Upload(*pdf)</label>
                                                     <div class="col-md-6">
-                                                        <input type="file" onchange="validateSize(this, 10240)"
-                                                            accept="application/pdf" id="assignFile" name="assignFile">
+                                                        <input type="file" onchange="validateSize(this, 10240)" accept="application/pdf" id="assignFile" name="assignFile">
                                                     </div>
                                                 </div>
                                                 <div class="row mb-3" hidden>
@@ -68,37 +58,32 @@
                                                     </div>
                                                 </div>
                                                 <div class="row mb-3">
-                                                    <label for="subject"
-                                                        class="col-md-4 col-form-label text-md-end">{{ __("Subject") }}</label>
+                                                    <label for="subject" class="col-md-4 col-form-label text-md-end">{{ __("Subject") }}</label>
                                                     <div class="col-md-6">
                                                         <input id="subject-ass" name="subject-ass" value="{{$subject->title}}" disabled>
                                                     </div>
                                                 </div>
                                                 <div class="row mb-3">
-                                                    <label for="title_ass"
-                                                        class="col-md-4 col-form-label text-md-end">{{ __("Title") }}</label>
+                                                    <label for="title_ass" class="col-md-4 col-form-label text-md-end">{{ __("Title") }}</label>
                                                     <div class="col-md-6">
                                                         <input id="title_ass" name="title_ass">
                                                     </div>
                                                 </div>
                                                 <div class="row mb-3">
-                                                    <label for="desc_ass"
-                                                        class="col-md-4 col-form-label text-md-end">{{ __("Description") }}</label>
+                                                    <label for="desc_ass" class="col-md-4 col-form-label text-md-end">{{ __("Description") }}</label>
                                                     <div class="col-md-6">
                                                         <input id="desc_ass" name="desc_ass">
                                                     </div>
                                                 </div>
                                                 <div class="row mb-3">
-                                                    <label for="due"
-                                                        class="col-md-4 col-form-label text-md-end">{{ __("Deadline") }}</label>
+                                                    <label for="due" class="col-md-4 col-form-label text-md-end">{{ __("Deadline") }}</label>
                                                     <div class="col-md-6">
                                                         <input id="due_ass" name="due_ass" type="date">
                                                     </div>
                                                 </div>
                                                 <div class="row mb-0">
                                                     <div class="col-md-6 offset-md-4">
-                                                        <button id="upload-assign" type="submit"
-                                                            class="btn btn-primary">
+                                                        <button id="upload-assign" type="submit" class="btn btn-primary">
                                                             {{ __('Submit') }}
                                                         </button>
                                                     </div>
@@ -110,7 +95,6 @@
                             </div>
                         </div>
                     </div>
-
                     @endif
                     @endforeach
                     @endforeach
@@ -140,15 +124,13 @@
                         <label class="col-md-4 col-form-label text-md-end" for="student">{{ __('Nama Judul')
                             }}</label>
                         <div class="col-md-6">
-                            <input onkeyup="symbolBoundaries('title')" id="title" name="title" class="form-control"
-                                required>
+                            <input onkeyup="symbolBoundaries('title')" id="title" name="title" class="form-control" required>
                         </div>
                     </div>
                     <div class="row mb-3">
                         <label for="formFile" class="col-md-4 col-form-label text-md-end">File Upload(*pdf)</label>
                         <div class="col-md-6">
-                            <input type="file" onchange="validateSize(this, 10240)" accept="application/pdf" id="file"
-                                name="file" class="file">
+                            <input type="file" onchange="validateSize(this, 10240)" accept="application/pdf" id="file" name="file" class="file">
                         </div>
                     </div>
                     <div class="row mb-3">
@@ -184,8 +166,8 @@
 </div>
 
 <script type="text/javascript">
-    $(document).ready(function () {
-        $('#upload').bind("click", function () {
+    $(document).ready(function() {
+        $('#upload').bind("click", function() {
             var txtTitle = $('#title').val();
             var imgVal = $('#file').val();
             var txtVal = $('#roomc').val();
@@ -195,7 +177,6 @@
             }
         });
     });
-
 </script>
 
 @endsection
